requirements-lock.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
altair==4.2.2
annotated-types==0.6.0
anyio==4.4.0
apache-airflow==2.8.4
apache-airflow-providers-common-compat==1.6.1
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.2
apache-airflow-providers-fab==1.2.2
apache-airflow-providers-ftp==3.10.0
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apispec==6.6.1
argcomplete==3.4.0
asgiref==3.8.1
attrs==23.2.0
Authlib==1.6.1
avro-python3==1.10.2
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-mgmt-authorization==3.0.0
azure-mgmt-compute==30.3.0
azure-mgmt-core==1.6.0
azure-mgmt-databricks==1.0.0
azure-mgmt-datafactory==2.0.0
azure-mgmt-eventhub==10.0.0
azure-mgmt-keyvault==10.2.0
azure-mgmt-monitor==5.0.0
azure-mgmt-msi==7.0.0
azure-mgmt-network==25.1.0
azure-mgmt-resource==24.0.0
azure-mgmt-sql==3.0.1
azure-mgmt-storage==21.0.0
azure-mgmt-synapse==2.0.0
azure-storage-blob==12.19.0
Babel==2.15.0
backoff==2.2.1
bandit==1.7.9
black==24.4.2
blinker==1.8.2
cachelib==0.9.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
coverage==7.10.4
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==41.0.7
delta-spark==3.2.0
Deprecated==1.2.14
detect-secrets==1.5.0
dill==0.3.1.1
distro==1.9.0
dnspython==2.6.1
docutils==0.16
dparse==0.6.4
email-validator==1.3.1
entrypoints==0.4
execnet==2.1.1
Faker==37.4.2
filelock==3.12.4
findspark==2.0.1
flake8==7.1.0
Flask==2.2.5
Flask-AppBuilder==4.3.11
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2023.12.2
google-re2==1.1.20240702
googleapis-common-protos==1.63.2
great_expectations==1.5.8
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.7.0
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.1.0
isodate==0.7.2
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kafka-python==2.2.15
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.5
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
mccabe==0.7.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
mistune==3.1.3
more-itertools==10.3.0
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.0.5
mypy==1.10.0
mypy_extensions==1.1.0
numpy==1.26.4
oauthlib==3.3.1
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
packaging==24.1
pandas==2.1.4
pathspec==0.12.1
pbr==6.1.1
pendulum==3.0.0
platformdirs==4.3.8
pluggy==1.5.0
posthog==6.4.1
prison==0.2.1
prometheus_client==0.22.1
propcache==0.3.2
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyarrow==21.0.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.2.3
pyspark==3.5.1
# Editable install with no version control (pyspark-interview-project==0.1.0)
-e /Users/kunal/IdeaProjects/pyspark_data_engineer_project
pytest==8.2.1
pytest-benchmark==4.0.0
pytest-cov==6.2.1
pytest-mock==3.14.1
pytest-spark==0.6.0
pytest-xdist==3.8.0
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.19.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.12.9
safety==3.2.4
safety-schemas==0.0.14
scipy==1.16.1
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
stevedore==5.4.1
structlog==25.4.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.2
toolz==1.0.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal-pathlib==0.1.4
urllib3==2.0.7
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.23.0
